354: Backend changes for pagination of child records fetch
