+++ src/controllers/children/index.ts
@@ -1,0 +1,5 @@
+export * from './changeEnrollment';
+export * from './create';
+export * from './delete';
+export * from './get';
+export * from './update';
