+++ src/controllers/users.ts
@@ -1,11 +1,29 @@
 import { User, Site, Organization } from '../entity';
 import { getManager } from 'typeorm';
 
+/**
+ * Augment a user object with site and organization data.
+ * Add all sites the user has access to, as well as any
+ * organizations they have access to OR organizations for
+ * the allowed sites (users will either have site OR organization
+ * level access, but not both)
+ * @param user
+ */
 export const addDataToUser = async (user: User) => {
-  const [sites, organizations] = await Promise.all([
-    getManager().findByIds(Site, user.siteIds || []),
-    getManager().findByIds(Organization, user.organizationIds || []),
-  ]);
+  const sites =
+    user.siteIds && user.siteIds.length
+      ? await getManager().findByIds(Site, user.siteIds)
+      : [];
+
+  //
+  const orgIds =
+    user.organizationIds && user.organizationIds.length
+      ? user.organizationIds
+      : sites.map((s) => s.organizationId);
+
+  const organizations = orgIds.length
+    ? await getManager().findByIds(Organization, orgIds)
+    : [];
 
   user.sites = sites;
   user.organizations = organizations;
