+++ src/routes/children.ts
@@ -11,25 +11,6 @@ import { parseQueryString } from '../utils/parseQueryString';
 export const childrenRouter = express.Router();
 
 /**
- * /children POST
- *
- * Creates a new child
- */
-childrenRouter.post(
-  '/',
-  passAsyncError(async (req, res) => {
-    try {
-      const child = await controller.createChild(req.body, req.user);
-      res.status(201).send({ id: child.id });
-    } catch (err) {
-      if (err instanceof ApiError) throw err;
-      console.error('Error creating child: ', err);
-      throw new BadRequestError('Child information not saved.');
-    }
-  })
-);
-
-/**
  * /children GET
  *
  * Returns all children the authed user has access to,
@@ -43,6 +24,8 @@ childrenRouter.get(
     }) as string[];
     const count = parseQueryString(req, 'count');
     const missingInfo = parseQueryString(req, 'missing-info') as string;
+    const skip = parseQueryString(req, 'skip', { post: parseInt }) as number;
+    const take = parseQueryString(req, 'take', { post: parseInt }) as number;
 
     if (count && count === 'true') {
       const count = await controller.getCount(req.user);
@@ -53,6 +36,8 @@ childrenRouter.get(
     const children = await controller.getChildren(req.user, {
       organizationIds,
       missingInfo,
+      skip,
+      take,
     });
 
     res.send(children);
@@ -101,6 +86,25 @@ childrenRouter.put(
     } catch (err) {
       if (err instanceof ApiError) throw err;
       console.error('Error saving changes to child: ', err);
+      throw new BadRequestError('Child information not saved.');
+    }
+  })
+);
+
+/**
+ * /children POST
+ *
+ * Creates a new child
+ */
+childrenRouter.post(
+  '/',
+  passAsyncError(async (req, res) => {
+    try {
+      const child = await controller.createChild(req.body, req.user);
+      res.status(201).send({ id: child.id });
+    } catch (err) {
+      if (err instanceof ApiError) throw err;
+      console.error('Error creating child: ', err);
       throw new BadRequestError('Child information not saved.');
     }
   })
